backend/requirements.txt
# Planet Code Forge Backend Dependencies

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
grpcio==1.60.1
grpcio-tools==1.60.1

# AI/ML & Code Analysis
scikit-learn==1.3.2
transformers==4.35.2
torch==2.1.2
numpy==1.24.4
pandas==2.1.4
nltk==3.8.1
tree-sitter==0.21.0

# Database & Caching
asyncpg==0.29.0
psycopg2-binary==2.9.9
redis==5.0.1
sqlalchemy==2.0.23
alembic==1.13.0

# Web3 & Blockchain
web3==6.13.0
eth-account==0.9.0
py-solc-x==2.0.2

# Security & Auth
cryptography==41.0.8
PyJWT==2.8.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Async & Streaming
aioredis==2.0.1
aiokafka==0.9.0

# Utilities
pydantic==2.5.2
python-dotenv==1.0.0
python-decouple==3.8
httpx==0.25.2
loguru==0.7.2
prometheus-client==0.19.0

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
flake8==6.1.0